requirements.txt
asgiref==3.3.4
boto3==1.17.62
botocore==1.20.62
Django==3.2
humanize==3.5.0
image==1.5.33
jmespath==0.10.0
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
